v1.0.0 - First Official Release

+++ Assets/Scripts/Data/BattleDefinitionSO.cs
@@ -7,6 +7,6 @@ public class BattleDefinitionSO : ScriptableObject
     public List<CharacterDataSO> players;
     public List<CharacterDataSO> enemies;
 
-    public int battleLevel;
+    public float battleLevel;
     public bool isWon;
 }
